voice-agent/requirements.txt
pvporcupine>=3.0.0
openwakeword>=0.6.0
faster-whisper>=1.0.0
PyAudio==0.2.14
numpy>=1.24.0
requests>=2.28.0
anthropic>=0.39.0
PyYAML>=6.0
python-dotenv>=1.0.0
pydub>=0.25.0
